clean up watch helper program
